refactored questions package
