--- src/main/java/dev/salonce/discordquizbot/core/questions/QuizOption.java
@@ -1,3 +1,0 @@
-package dev.salonce.discordquizbot.core.questions;
-
-public record QuizOption(String text, boolean isCorrect) {}
