+++ src/main/java/dev/salonce/discordquizbot/core/questions/questions/Question.java
@@ -1,4 +1,4 @@
-package dev.salonce.discordquizbot.core.questions;
+package dev.salonce.discordquizbot.core.questions.questions;
 
 import dev.salonce.discordquizbot.core.questions.rawquestions.RawQuestion;
 import lombok.Getter;
