+++ src/main/java/dev/salonce/discordquizbot/core/questions/questions/QuizOption.java
@@ -1,0 +1,3 @@
+package dev.salonce.discordquizbot.core.questions.questions;
+
+public record QuizOption(String text, boolean isCorrect) {}
