+++ src/main/java/dev/salonce/discordquizbot/core/questions/topics/DifficultyLevel.java
@@ -1,4 +1,4 @@
-package dev.salonce.discordquizbot.core.questions;
+package dev.salonce.discordquizbot.core.questions.topics;
 
 import dev.salonce.discordquizbot.core.questions.rawquestions.RawQuestion;
 import lombok.Getter;
@@ -8,7 +8,7 @@ import java.util.List;
 
 @Getter
 class DifficultyLevel {
-    private List<RawQuestion> rawQuestions = new ArrayList<>();
+    public List<RawQuestion> rawQuestions = new ArrayList<>();
 
     public DifficultyLevel(List<RawQuestion> rawQuestions) {
         addRawQuestions(rawQuestions);
